Make compiler_invoke a method on CompileStruct
